Update phpadminer.php
phpMyAdmin is one of the most famous tools for managing the MySQL database. However, it is not perfect which is the reason why Adminer has been created. In which areas is Adminer a better alternative to replace phpMyAdmin? The list is quite long and some points may be insignificant for you. The most important differences are:

User experience
Supported MySQL features
Performance
Security
There are several screenshots linked from the comparison table further explaining the difference.